@@ -1,4 +1,4 @@
-name: public_tools
+name: shortcut_launcher
 description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
@@ -18,19 +18,18 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  sqflite: ^1.3.2+2
-  path_provider: ^2.0.8
-  shared_preferences: ^2.0.11
+  sqflite: ^2.0.2
+  path_provider: ^2.0.9
+  shared_preferences: ^2.0.13
   oktoast: ^3.1.5
   intl: ^0.17.0
-  cached_network_image: ^2.5.1
+  cached_network_image: ^3.2.0
   flutter_highlight: ^0.7.0
   hotkey_manager: ^0.1.6
-  flutter_js: 0.5.0+5
-  window_manager: ^0.1.1
+  window_manager: ^0.2.0
   logger: ^1.1.0
   flutter:
     sdk: flutter
@@ -38,13 +37,13 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.4
   flutter_html: ^2.2.1
-  provider: ^6.0.2
   uuid: ^3.0.6
+  collection: ^1.15.0-nullsafety.4
 
 dev_dependencies:
-  pigeon: ^1.0.12
+  pigeon: ^2.0.3
   flutter_test:
     sdk: flutter
 
