@@ -1,43 +1,41 @@
 import 'dart:async';
 import 'dart:convert';
-import 'dart:ffi';
 import 'dart:io';
 
-import 'package:oktoast/oktoast.dart';
-import 'package:public_tools/core/plugin.dart';
-import 'package:public_tools/pigeon/app.dart';
-import 'package:public_tools/utils/logger.dart';
 import 'package:uuid/uuid.dart';
-import 'package:window_manager/window_manager.dart';
 
-typedef void ListReceiver(List<BaseListItem> list);
-typedef void EnterItemReceiver();
+import '../../utils/logger.dart';
 
-List<WebSocket> sockets = [];
-List<ListReceiver> receivers = [];
-List<EnterItemReceiver> enterItemReceivers = [];
-void Function(String content) setResultItemPreview = (content) => null;
-String _binPath = '';
+Map<String, dynamic> _addToMap(
+    String key, dynamic value, Map<String, dynamic> map) {
+  map[key] = value;
+  return map;
+}
 
 class MessageData {
-  String type;
+  String? type;
   Map<String, dynamic> payload;
   int time;
   String id;
   String replyId;
 
-  MessageData({this.type, this.payload})
-      : time = DateTime.now().millisecondsSinceEpoch,
-        id = Uuid().v4(),
-        replyId = Uuid().v4();
+  MessageData({
+    this.type,
+    required this.payload,
+    int? time,
+    String? id,
+    String? replyId,
+  })  : time = time ?? DateTime.now().millisecondsSinceEpoch,
+        id = id ?? Uuid().v4(),
+        replyId = id ?? Uuid().v4();
 
-  MessageData.fromJson(Map<String, dynamic> json) {
-    type = json['type'];
-    payload = json['payload'];
-    time = json['time'];
-    id = json['id'];
-    replyId = json['replayId'];
-  }
+  MessageData.fromJson(Map<String, dynamic> json)
+      : type = json['type'],
+        payload = json['payload'],
+        time = json['time'],
+        id = json['id'],
+        replyId = json['replyId'];
+
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
     data['type'] = this.type;
@@ -47,85 +45,84 @@ class MessageData {
     data['time'] = this.time;
     return data;
   }
+
+  MessageData makeReplyMessage(Map<String, dynamic> payload) {
+    return MessageData(
+      type: this.type,
+      payload: payload,
+      id: this.replyId,
+    );
+  }
+
+  MessageData.makeEventMessage(String eventName, Map<String, dynamic> payload)
+      : this.payload = _addToMap('event', eventName, payload),
+        this.type = 'event',
+        this.id = Uuid().v4(),
+        this.replyId = Uuid().v4(),
+        this.time = DateTime.now().millisecondsSinceEpoch;
 }
 
-String makeMessageData(String type, Map<String, dynamic> payload) {
+String makeMessageData(String? type, Map<String, dynamic> payload) {
   return jsonEncode(MessageData(type: type, payload: payload).toJson());
 }
 
-void Function(dynamic) _createHandler(WebSocket socket) {
-  return (dynamic data) async {
-    final message = MessageData.fromJson(data);
-    logger.i(message.toJson());
-    if (message.type == 'list') {
-      List<BaseListItem> list = message.payload["list"]
-          .map<BaseListItem>((e) => BaseListItem.fromJson(e))
-          .toList();
-      receivers.forEach((element) {
-        element(list);
-      });
-    } else if (message.type == 'enter') {
-      final enterItemReceiver = enterItemReceivers.first;
-      if (enterItemReceiver != null) {
-        enterItemReceiver();
-      }
-    } else if (message.type == 'toast') {
-      showToast(message.payload["content"]);
-    } else if (message.type == 'hideApp') {
-      await Service().hideApp();
-    } else if (message.type == 'showApp') {
-      await windowManager.show();
-    } else if (message.type == 'preview') {
-      setResultItemPreview(message.payload['html']);
-    }
-    // socket.add(jsonEncode(data));
-  };
-}
-
 class RemotePluginServer {
-  HttpServer _server;
-  WebSocket _socket;
-  final Function onReady;
-  Map<String, StreamController<MessageData>> _messages = {};
+  HttpServer? _server;
+  WebSocket? _socket;
+  Map<String, List<void Function(Map<String, dynamic>)>> _handlers = {};
 
-  RemotePluginServer({this.onReady}) {
+  final Function onReady;
+  final Function onDisconnect;
+  Map<String?, StreamController<MessageData>> _messages = {};
+
+  RemotePluginServer({required this.onReady, required this.onDisconnect}) {
     this._init();
   }
 
   void _init() async {
     this._server = await HttpServer.bind('127.0.0.1', 4040);
-    this._server.listen((HttpRequest req) async {
-      if (req.uri.path != '/ws' || this._socket != null) {
-        return;
-      }
-      logger.i('新的socket已连接');
-      var socket = await WebSocketTransformer.upgrade(req);
-      socket.map((string) => jsonDecode(string)).listen(_onMessage, onDone: () {
-        socket.close();
-        this._socket = null;
-        runClient();
-      });
-      this._socket = socket;
-      this.onReady();
-    });
+    this._server!.listen(
+      (HttpRequest req) async {
+        logger.i('新的链接: ${req.uri}');
+        if (req.uri.path != '/ws' || this._socket != null) {
+          return;
+        }
+        logger.i('新的socket已连接');
+        // ignore: close_sinks
+        var socket = await WebSocketTransformer.upgrade(req);
+        socket.done.then((data) {
+          logger.i('socket done');
+          this._socket!.close();
+          this._socket = null;
+          this.onDisconnect();
+        });
+        socket.handleError((error) {
+          logger.i('socket error: $error');
+          this._socket!.close();
+          this._socket = null;
+          this.onDisconnect();
+        });
+        socket.map((string) => jsonDecode(string)).listen(_onMessage);
+        this._socket = socket;
+        this.onReady();
+      },
+    );
   }
 
   void _onMessage(dynamic data) async {
     final message = MessageData.fromJson(data);
     logger.i('收到消息: ${message.toJson()}');
     if (_messages[message.id] != null) {
-      _messages[message.id].add(message);
-      _messages[message.id].close();
+      _messages[message.id]!.add(message);
+      _messages[message.id]!.close();
       _messages.remove(message.id);
     }
-    if (message.type == 'toast') {
-      showToast(message.payload["content"]);
-    } else if (message.type == 'hideApp') {
-      await Service().hideApp();
-    } else if (message.type == 'showApp') {
-      await windowManager.show();
-    } else if (message.type == 'preview') {
-      setResultItemPreview(message.payload['html']);
+    if (message.type == 'event') {
+      if (_handlers[message.payload['event']] != null) {
+        _handlers[message.payload['event']]!.forEach((element) {
+          element(message.payload);
+        });
+      }
     }
   }
 
@@ -136,15 +133,15 @@ class RemotePluginServer {
     // ignore: close_sinks
     final streamController = StreamController<MessageData>();
     _messages[message.replyId] = streamController;
-    /* Future.delayed(Duration(seconds: 5), () {
+    Future.delayed(Duration(seconds: 5), () {
       if (_messages[message.replyId] != null) {
-        _messages[message.replyId]
+        _messages[message.replyId]!
             .add(MessageData(type: 'error', payload: {'message': 'timeout'}));
-        _messages[message.replyId].close();
+        _messages[message.replyId]!.close();
         _messages.remove(message.replyId);
       }
-    }); */
-    _socket.add(json.encode(message));
+    });
+    _socket!.add(json.encode(message));
     final result = await streamController.stream.first;
     if (result.type == 'error') {
       throw Exception(result.payload['message']);
@@ -152,33 +149,29 @@ class RemotePluginServer {
     _messages.remove(message.time);
     return result.payload;
   }
+
+  void on(eventName, void Function(Map<String, dynamic>) callback) {
+    final eventHandlers = _handlers[eventName] ??= [];
+    eventHandlers.add(callback);
+    _handlers[eventName] = eventHandlers;
+  }
+
+  void off(eventName, void Function(Map<String, dynamic>) callback) {
+    final eventHandlers = _handlers[eventName] ??= [];
+    eventHandlers.remove(callback);
+    _handlers[eventName] = eventHandlers;
+  }
 }
 
-void setBinPath(String binPath) {
-  _binPath = binPath;
-}
-
-void runClient() async {
+void runClient(binPath) async {
   var pluginDir = '';
   if (Platform.environment["REMOTE_PLUGIN_MODE"] == 'local') {
     pluginDir = Directory.current.path + '/plugins';
   }
-  // @todo 生产环境下载并执行
-  if (pluginDir == '') return null;
   final clientProcess = await Process.start(
-    '$_binPath/npm',
+    '$binPath/npm',
     ['run', 'start'],
     workingDirectory: pluginDir,
   );
   clientProcess.stdout.transform(utf8.decoder).forEach(print);
-}
-
-void send([String type, Map<String, dynamic> payload = const {}]) {
-  sockets.forEach((element) {
-    element.add(makeMessageData(type, payload));
-  });
-}
-
-void onUpdateList(ListReceiver receiver) {
-  receivers.add(receiver);
 }
