@@ -4,13 +4,13 @@ import '../models/PasteItem.dart';
 
 class Utils {
   // 工厂模式
-  factory Utils() => _getInstance();
-  static Utils get instance => _getInstance();
-  static Utils _instance;
+  factory Utils() => _getInstance()!;
+  static Utils? get instance => _getInstance();
+  static Utils? _instance;
   Utils._internal() {
     // 初始化
   }
-  static Utils _getInstance() {
+  static Utils? _getInstance() {
     if (_instance == null) {
       _instance = new Utils._internal();
     }
@@ -34,9 +34,9 @@ class Utils {
 }
 
 // 只调用一次的函数
-T Function(A a, B b) once<A, B, T>(T Function(A a, B b) func) {
+T? Function(A a, B b) once<A, B, T>(T Function(A a, B b) func) {
   var alreadyRun = false;
-  T t;
+  T? t;
   return (A a, B b) {
     if (alreadyRun) return t;
     return func(a, b);
@@ -44,9 +44,9 @@ T Function(A a, B b) once<A, B, T>(T Function(A a, B b) func) {
 }
 
 // 可取消的函数
-T Function(A a, B b) cancelable<A, B, T>(T Function(A a, B b) func) {
+T? Function(A a, B b) cancelable<A, B, T>(T Function(A a, B b) func) {
   bool canceled = false;
-  T t;
+  T? t;
   var canceledFn = (A a, B b) {
     if (!canceled) {
       return func(a, b);
@@ -55,3 +55,5 @@ T Function(A a, B b) cancelable<A, B, T>(T Function(A a, B b) func) {
   };
   return canceledFn;
 }
+
+// latest future
