@@ -1,10 +1,10 @@
 import 'package:pigeon/pigeon.dart';
 
 class InstalledApplication {
-  String name;
-  String path;
-  String icon;
-  String pinyin;
+  String? name;
+  String? path;
+  String? icon;
+  String? pinyin;
 }
 
 // Flutter 调用原生代码
