@@ -1,9 +1,9 @@
-// const mdnPlugin = require('./mdn')
+const mdnPlugin = require('./mdn')
 const qrcodePlugin = require('./qrcode/index.js')
 // const magicPlugin = require('./magic')
-// const translatePlugin = require('./translate')
-// const pluginStorePlugin = require('./store')
+const translatePlugin = require('./translate')
+const storePlugin = require('./store')
 
-module.exports = () => ([qrcodePlugin])
+module.exports = () => ([qrcodePlugin, mdnPlugin, storePlugin, translatePlugin])
 
 console.log('remote plugin is running')
